+++ custom_components/ha_mcp_bridge/const.py
@@ -5,9 +5,13 @@ DOMAIN = "ha_mcp_bridge"
 CONF_HOST = "host"
 CONF_PORT = "port"
 CONF_UPDATE_INTERVAL = "update_interval"
+CONF_MCP_URL = "mcp_url"
+CONF_MCP_TOKEN = "mcp_token"
 
 DEFAULT_HOST = "http://127.0.0.1"
 DEFAULT_PORT = 8099
 DEFAULT_HEALTH_PATH = "/health"
 DEFAULT_STATUS_PATH = "/status"
 DEFAULT_UPDATE_INTERVAL = 60
+DEFAULT_MCP_URL = ""
+DEFAULT_MCP_TOKEN = ""